moved code for calculating optimalsize to utils method setupOptimalCameraPreviewSize
